requirements.txt
crewai
langchain
google-generativeai
duckduckgo-search
pydantic
python-dotenv
tqdm
pandas
rich
requests
wikipedia
streamlit